Fix server restart failure with binary log enabled
Summary:
During startup, the data dictionary scans through a set of keys to find
the table names. These keys use the index prefix, which is 4 bytes. The
scanner knows it has hit the end of the list when the key it is
examining no longer matches the index prefix.

However, the scanner first verifies the length of the key is larger than
the index prefix before determining if the key is a data dictionary key.
If this key is not a data dictionary, but is 4 bytes (like the binlog
key), then it returns a failure and stops initialization.

Fix is to determine if the key has the index prefix before failing it
for invalid length.

Differential Revision: https://reviews.facebook.net/D32481

fbshipit-source-id: 7b15994de1c