Disable logical read-ahead when transaction is switching between tables
Summary:
WebScaleSQL Feature: Logical Read-Ahead

Every time a transaction starts to scan a new table, logical read-ahead
will clean all the previous buffers and start prefetching on the new
index from the beginning. This doesn't work well when a transaction is
switching frequently among many tables, which happens for partitioned
tables in many cases. This change disables LRA when too many tables are
involved in a transaction. A new session variable, innodb_lra_n_spaces,
is introduced to set the maximum number of table access switches allowed
before LRA is disabled.

The default value of this parameter is 3 (That is, logical read-ahead
will be disabled if it is accessing different tables 3 times in a row).

Note: This diff depends on these other diffs:

https://reviews.facebook.net/D23463
https://reviews.facebook.net/D23469

Reference Patch: https://github.com/facebook/mysql-5.6/commit/6752347

Differential Revision: D6686690 (https://github.com/facebook/mysql-5.6/commit/2bcb9aec021be8993341463f0b1bd6fc5fba38eb)

fbshipit-source-id: 8156f78c84c