Abort mysqld if failing to write to WAL
Summary:
Depending on kernel and storage, but there is a possibility that
writing to binlog succeeds (at binlog commit) but writing to
WAL fails (at rocksdb commit). Currently commit() silently ignores
WAL write errors at flush_batch_intern(). This means clients
get "OK" reply from mysqld even though mysqld fails to write to WAL.
This is very dangerous, because transactions are written to binlog
and replicated but not persisted in MyRocks.

This diff checks return code from DBImpl::Write(), and mysqld aborts
if not getting OK. This is much safer behavior because clients get
errors if failing to write.

Differential Revision: https://reviews.facebook.net/D37707

fbshipit-source-id: 83c295203ef