Replace cf options file with MySQL system variable
Summary:
Currently to override the default settings for a particular column family you have to create a
separate file and point a MySQL variable (rocksdb_cf_options_file) at the file.  It would be simpler to
have a MySQL system variable that allows you to override the settings on a per-column family basis.  The
new variable is rocksdb_override_cf_options and is in the format <cf>={<options}[;<cf>={<options>}]...

Each column family should only be specified once (it would be an error to include it again) and the options
must be surrounded by curly braces.  The options inside follow the format <key>=<value>[;<key>=<value>]...
like in rocksdb_default_cf_options and will be validated by RocksDB.

Differential Revision: https://reviews.facebook.net/D49293

fbshipit-source-id: d6b4ae50b5d