Improve singled thread replication performance
Summary:
On slave we update the slave_gtid_info table through MySQL. Delegate this
to storage engine which is more efficient. SQL threads updates the row information
in THD structure. Storag engines (InnoDB and Rocksdb) uses this information and
updates the slave_gtid_info table during prepare phase.

Note the table's schema is not modified and it is only accessed by SQL threads.

Differential Revision: https://reviews.facebook.net/D63033

fbshipit-source-id: 8054a43ab14