Fix the incorrect operator new usage under storage/rocksdb
Summary:
Based on the discussions related to OOM handling in MySQL I observed that currently there are some instances of `new` operator usage which are incorrect. We compile our code with exceptions enabled (requirement implied by the fact that we use RocksDB and STL). This means that when `new` operator fails to allocate requested amount of memory then it'll throw `std::bad_alloc`. It won't return `NULL` (we don't use `std::nothrow` so this case is taken care of as well).

Current code needs to be fixed for the following reasons:

- The error handling code won't be executed in case of OOM and therefore there's no point in having it.

- Reading the code implies that we'll do something, //e.g.//, return an error code when OOM happens. That creates a false sense of security and enables people to propagate this incorrect pattern.

Reviewed By: jtolmer

Differential Revision: https://reviews.facebook.net/D53307

fbshipit-source-id: 9c71357332e