Use same snapshot for iterator as for Get/GetForUpdate on index scans
Summary:
When using an index scan we were not specifically creating a snapshot until the Get or GetForUpdate call, but the iterator was using an implicit snapshot.  This caused us to 'see' rows that were not there and other problems.  This diff changes this so we use the same snapshot for the iterator and the Get/GetForUpdate call.

In addition I've disabled some tests that we know will fail (until we add a retry option) and added test cases for some of the concurrent access cases we've identified

Differential Revision: https://reviews.facebook.net/D54129

fbshipit-source-id: ff7cc4fcfce