Add TSan build support for internal builds
Summary:
Adding TSan support is tricky because requirements aren't clearly documented and behavior appears to vary between CentOS and Ubuntu for example. Add to the mix different versions of libraries and differences in build environments and you will get a nontrivial problem to solve.

This is first batch of the TSan work. What will follow is:

  - Internal changes to enable the TSan builds.
  - More external changes to support TSan outside FB build environment (as verified to be working on Ubuntu).

Reviewed By: jtolmer

Differential Revision: https://reviews.facebook.net/D56259

fbshipit-source-id: f936af1d15e