Make building unittest directory conditional on WITH_EMBEDDED_SERVER
Summary:
This is the fix for https://github.com/facebook/mysql-5.6/issues/147. A number of individuals using MyRocks externally have encountered this problem. We don't see it internally because we build with a different set of options.

The key here is to make sure that we check if `WITH_EMBEDDED_SERVER` is enabled or not.

Reviewed By: vasilep

Differential Revision: https://reviews.facebook.net/D56283

fbshipit-source-id: 5bd4f16816c