Add an option to skip using bloom filter for reads in MyRocks
Summary:
This is a debugging feature to optionally skip to use bloom filter for
reads, except point lookup (Get and GetForUpdate). Bloom filter is still
created for new SST files, but Seek() does not use bloom filter anymore
with this option. This can be used for debugging bloom filter bugs.

Differential Revision: https://reviews.facebook.net/D56523

fbshipit-source-id: ad5ca3c18d7