fixed crash with fast drop index and rocksdb_strict_collation exceptions
Summary:
If in the context of in-place alter table, table name from the altered
TABLE object (table_arg) will be a temp table (#sql-xxxx format).
To check collation support we need the original table name, so if a
temporary table is found we extract the name from the db_table parameter
rather than the altered TABLE.

Fixes https://github.com/facebook/mysql-5.6/issues/270.

Differential Revision: https://reviews.facebook.net/D59997

fbshipit-source-id: b08769e4278