Allow tests to be skipped if full regex support is missing (#399)
Summary: GCC 4.8 did not support full regex but did include the functions.  If we try to run some of our tests that intentionally cause failures the system crashes when built with GCC 4.8.  Add the ability to detect what compiler and version the system was built with so we can skip tests that are known to fail without full regex support

Reviewed By: lloyd

Differential Revision: D4134960 (https://github.com/facebook/mysql-5.6/commit/8b8334bf00e51692bdbd83b5a010601b6b903699)

fbshipit-source-id: 152b8382940