add explicit fsync to fast secondary index creation
Summary:
Due to the large amount of data written to the os due to fast secondary index
creation, we need to add explicit fsyncs after each call to my_write, as
my_write does not guarantee the data to be flushed to disk.  Otherwise,
potentially writeback and page cache stalls may happen when several indexes are
being created concurrently.

Reviewed By: gunnarku

Differential Revision: D4330649 (https://github.com/facebook/mysql-5.6/commit/f91e2ffd89a994fb79f1f88551f0bf4ac86bc020)

fbshipit-source-id: 9cf216cd383