Stop slave if master's binlog format is not ROW but some features that depend on it are enabled on the slave
Summary:
Idempotent recovery and rocksdb's skip trx api should be enabled only
when master's binlog format is ROW. We detect master's binlog format by
checking if it executes any DML statements. This tells us that the binlog
format is not ROW and we stop the slave.
Also, I moved rpl_skip_tx_api from rocksdb plugin to global mysqld so
that it can be retrieved without using a handler.

Reviewed By: tianx

Differential Revision: D4485384

fbshipit-source-id: 1adf70cf5a9