fix cardinality for secondary keys created via inplace alter (#531) (#531)
Summary:
Cardinality requires the keydef during the sst file creation to be calculated. Without the keydef, the collector doesn't store cardinality stats. This adds a member to the ddl_manager so that the properties collector can find the keydef ongoing creation before it is persisted to the data dictionary.
Closes https://github.com/facebook/mysql-5.6/pull/531

Differential Revision: D4563694 (https://github.com/facebook/mysql-5.6/commit/5d2b9539f6319f56e81165797dfa0ea0f8e0126a)

Pulled By: alxyang

fbshipit-source-id: 63c1d771e53