Add test coverage for partitioning support in MyRocks
Summary:
To make sure that we have a good baseline for partitioning support in MyRocks, we need to have a nontrivial amount of test coverage. This set of changes uses examples defined in https://dev.mysql.com/doc/refman/5.6/en/partitioning.html to validate that MyRocks behaves as expected for a variety of partitioning supported scenarios.

There are no code changes here, just test definitions and expected results.

Differential Revision: D4605936 (https://github.com/facebook/mysql-5.6/commit/fd8364645f2a16c2ac6b0ba4a16816cf8ac4ff10)

fbshipit-source-id: 0cc2c8abc19