Support dynamic CF updates using `rocksdb_update_cf_options`
Summary:
We want to be able to update CF options dynamically during the runtime. To do so a `rocksdb_update_cf_options` system variable is introduced. It uses the same format as `rocksdb_override_cf_options` and enables specifying a CF and its properties. If parsing new options succeeds then we call `SetOptions()` to update CF options.

Error handling and validation is left to RocksDB. If dynamic update of a particular option is not supported then we log a warning message about the failed operation. In case of success `ROCKSDB_CF_OPTIONS` table will reflect the latest values.

Reviewed By: lloyd

Differential Revision: D4773976 (https://github.com/facebook/mysql-5.6/commit/a5588b88f2ff9badda5077667289cf97299a1aa2)

fbshipit-source-id: f94d1344019