Enter into a crash loop on detected data corruption
Summary:
The main idea of this branch is to keep server crashing during restarts
once rocksdb reported data corruption.

The other big change is to start check status of iterator. If the
iterator goes to corrupted entry it becomes invalid and could look like
EOF if status is not checked. This branch introduces this check and
aborts the server if it finds corrupted value.

One more case when an error could be surfaced is background operation.
This branch starts logging this errors and abort if error is corruption.

Once corruption is detected server writes marker file (named
ROCKSDB_CORRUPTED) in the data directory and aborts. If marker files
exists on server then process exists normally with error message so that mysqld_safe
won't restart it. The restart failure can be switched of by starting mysqld with --rocksdb_allow_to_start_after_corruption=1 flag.

Differential Revision: D5983297 (https://github.com/facebook/mysql-5.6/commit/c28e1997cff502147cf189ea1d71efdad44c9ee0)

fbshipit-source-id: 983d450ad78