Skip trailing space bytes for non-unpackable fields (#785) (#785)
Summary:
We always generate unpack info bytes that indicate how much trailing space we need, even when the field can never be unpacked, because it is not implemented. When skipping these fields however, we only skip the key portion, and not the value portion which corrupts data for code that reads unpack_info downstream. Fix by skipping these bytes.

Closes #784
Closes https://github.com/facebook/mysql-5.6/pull/785

Differential Revision: D6699483 (https://github.com/facebook/mysql-5.6/commit/f3d36951a62aa1a00a5a10083fb77e9919022e03)

fbshipit-source-id: 7707a3173f7