Add query attributes
Summary:
Currently, queries sent over the network are encoded as follows:
  total packet length
  COM_QUERY
  query string

The length of the query string is implicit from and calculated from the packet length. This change adds a new command called COM_QUERY_ATTRS with the following structure
  total packet length
  COM_QUERY_ATTRS
  total attributes length
  attributes string
  query string

Note that query string length is still implicit. In order to avoid conflicting RPC values with Oracle, COM_QUERY_ATTRS was put at the end of the valid RPC range. Its value is set at 255.

The attributes are stored on the THD object.

To support testing, the client libraries were extended to include three new options so that mysqltest can use them:
  MYSQL_OPT_QUERY_ATTR_RESET
  MYSQL_OPT_QUERY_ATTR_ADD
  MYSQL_OPT_QUERY_ATTR_DELETE
which would reset, add or delete key value pairs from a set of query attributes held in the MYSQL handle on the client. If the client library sees that the set of query attributes is non-empty, it will send a COM_QUERY_ATTRS instead of a regular COM_QUERY.

The mysqltest client is extended with 3 new commands corresponding to the 3 options.
  query_attrs_reset
  query_attrs_add <key> <value>
  query_attrs_delete <key>
Reset takes no arguments and clears all query attributes currently stored. Add takes two arguments, the key and value, and stores them in the set of attributes. Delete takes one argument, the key, and deletes the pair in the set of attributes with the corresponding key.

Reference patch: https://github.com/facebook/mysql-5.6/commit/a619992

Differential Revision: D6916940 (https://github.com/facebook/mysql-5.6/commit/44063feef20eeffdab80f21d62fc31b8179a37be)

fbshipit-source-id: 9ffadcbf1bb