Prevent cert/key mismatch is combined file replaced between cert load and key load
Summary:
MySQL was using the functions that take a file as input for setting the
ssl cert and key.
There is a race condition if the cert file gets replaced between the
moment the cert is loaded and the key is loaded.
Fix: open file once and load cert and key. Used BIO functions that work
on file descriptors.

Reference patch: https://github.com/facebook/mysql-5.6/commit/53d0673

Differential Revision: D7299948 (https://github.com/facebook/mysql-5.6/commit/2203bcc585a4a93989df2e913f539704f7732de7)

fbshipit-source-id: cf2d1c134b5