Add option to skip big tests
Summary:
Our tests are timing out because we're running big tests even when we did not ask for them.

This is because as of https://github.com/mysql/mysql-server/commit/f0bcecb315a2, mtr will now turn on disabled and big tests even if the user did not ask for them. Add an option to override this behaviour.

Differential Revision: D14115272 (https://github.com/facebook/mysql-5.6/commit/b3caf3493e2c0c6dbea254a4dfa3b3d46a8ea83a)

fbshipit-source-id: 0dbe4911846