FB8-172: Add variable to disable full table/index scans (#970) (#970)
Summary:
Jira issue: https://jira.percona.com/browse/FB8-172

Reference Patch: https://github.com/facebook/mysql-5.6/commit/b90e801
Reference Patch: https://github.com/facebook/mysql-5.6/commit/ecef4b2

There are cases where it is always unacceptable for a client to be doing full table scans. To prevent this from happening, add a new variable optimizer_full_scan. When it is turned off, it will return a new error code ER_FULL_SCAN_DISABLED.

For best results, use with optimizer_force_index_for_range. Currently, it does not fallback to alternative plans because we check at the end of query planning.

EXPLAIN queries with derived tables do not populate select_options in the JOIN structure properly. This means that we could error if we tried to run an explain on a query with derived tables.

Instead of reading from the JOIN struct, read from the LEX struct on the THD, as this is where the flag is original set on during query parsing.

This bug also means that in upstream, we increment some status variables tracking full table scans despite the fact that only an explain statement was done. This seems to have been fixed in 8.0 though.

Pull Request resolved: https://github.com/facebook/mysql-5.6/pull/970

Reviewed By: lloyd

Differential Revision: D14567846 (https://github.com/facebook/mysql-5.6/commit/707435385a820e1df294e36daf3aaa7973cdd7d8)

Pulled By: lth

fbshipit-source-id: f7a7ed38322