PSI fix
Summary:
This is part of MyRocks 8.0 prep work - front loading as much work as possible in 5.6 before we actually port MyRocks to 8.0.

In this case, MySQL 8 perfscheme has changed quite a bit and we should've used the common wrapper for registering mutex/rwlock anyway. The if check is unnecessary as MySQL's own code doesn't seem to do it anyways.

This is a port of https://github.com/facebook/mysql-5.6/pull/930/commits/7b851f3b79383cf2d9c60a41c2105c9a5f6fcb76.

Reviewed By: lloyd

Differential Revision: D15302656 (https://github.com/facebook/mysql-5.6/commit/0faa001c53b4c8ad6fe41753c3b96fcee75ce91e)

fbshipit-source-id: a9fed9e6c18