Use DBUG_OFF instead of NDEBUG
Summary:
This is a port of https://github.com/facebook/mysql-5.6/pull/930/commits/9801160c0ade2e34ea30d4e0809789d30f7a1271

I'm not yet fully convinced the fix is required in 8.0 but given that most of other changes from georgelorchpercona are pretty solid anyways and this change is good in principle so I'm porting this anyway.

I did change the #if !defined to #ifndef to align with rest of the code.

Reviewed By: lloyd

Differential Revision: D15307412 (https://github.com/facebook/mysql-5.6/commit/06e1d3aea4522eaef1ce26f35591efed0a529a7d)

fbshipit-source-id: e2ef3abf690