Fix rocksdb.group_min_max
Summary:
We had a magic system variable `rocksdb_debug_optimizer_n_rows` that can "stabilize" the view of optimizer, but it only kicks in when stats.record == 0, which won't be true once ANALYZE TABLE (ha_rocksdb::analyze) is called and the row stats get updated. As a result, the optimizer view remains unstable and this leads to unstable tests.

The fix is to always overwrite the row stats if `rocksdb_debug_optimizer_n_rows` is specified.

Reviewed By: lloyd

Differential Revision: D15347479 (https://github.com/facebook/mysql-5.6/commit/75929903de7e6d02865d671d95e9f7eec6f0c88d)

fbshipit-source-id: aadb1e7192a