Properly support SSL options in mysqlslap
Summary: Mysqlslap doesn't apply options the same way for the initial connection and the workers - as a result the workers missing a lot of options passed from command line, including SSL. This fixes the SSL part only. I'll address the bigger issue with probably a bit refactoring with a separate fix (which is going to require quite a bit more testing).

Differential Revision: D16015149 (https://github.com/facebook/mysql-5.6/commit/ff6cc9f5dd9e7e913e804fcd9ab33895a49729c2)

fbshipit-source-id: 3c25da7b5fc