Fix mysqldump handling of varbinary columns
Summary:
https://bugs.mysql.com/bug.php?id=96053

8.0.13 added _binary tags to blobs in the mysqldump output. However,
memory allocated for the insert statement did not account of the 8 bytes
of the "_binary " string properly. This resulted in mysqldump writing
to memory beyond the buffer allocated for the string buffer before it
is reallocated with a new larger size.

The fix is to account for the 8 bytes of the "_binary " string during
the realloc calculation.

Reviewed By: lloyd

Differential Revision: D16088056 (https://github.com/facebook/mysql-5.6/commit/5e15108bcf25e811af70a688031ff5597e30e22e)

fbshipit-source-id: 7d8cd4398f1