Fixed failing tests rocksdb.cons_snapshot_read_committed and rocksdb.cons_snapshot_repeatable_read.
Summary:
Tests rocksdb.cons_snapshot_read_committed and rocksdb.cons_snapshot_repeatable_read are failing intermittently.
The direct cause of the failure is that number of snapshots doesn't match. With write_prepared compaction takes a snapshot. Depending on if background compaction is running or not the count of snapshot can be different.
The early fix D13960645 (https://github.com/facebook/mysql-5.6/commit/3115eacfce5551281497fb35bfd2f0102ffecda8) to disable auto compactions seems not working. It sets rocksdb_default_cf_options in opt files. But rocksdb_default_cf_options is a read-only system variable and it is not set during the test run.

The fix is to set disable_auto_compactions option to true for default and system column families (the tests only use these 2 column families) at the beginning of the tests and restore them afterwards.

Differential Revision: D17144475 (https://github.com/facebook/mysql-5.6/commit/6bff4a90c1c02e660d3a1ffd1bc3d1bc3c242ec8)

fbshipit-source-id: 2ec408fd622