Add an option to mysqlbinlog to abort if invalid events are found on in-use binlogs
Summary:
If mysqlbinlog finds a corrupted binlog event (e.g. binlog event checksum failure), it aborts if the binlog file was properly closed. However, it silently exits with exit code 0 if the corrupted event was found on the active (marked as LOG_EVENT_BINLOG_IN_USE_F) binlog files.
https://github.com/facebook/mysql-5.6/commit/ce2f22a8bdd78895d796087b3d09e244a324358f

But in some cases, we may want to know if binlog files have corrupted events on in-use (not properly closed) binlog files as well, rather than silently ignoring errors.

Adding a new command line argument "skip-verify-if-open" to control whether to ignore event errors on in-use binlog files.

Differential Revision: D17884872 (https://github.com/facebook/mysql-5.6/commit/366dcc292c8e8a7be6701f88b082a2d2a7c60cdc)

fbshipit-source-id: fc90b428beb