Start HA_ERR_ROCKSDB_FIRST at 500 and fix overlapping test error codes
Summary:
Given that we code HA_ERR_ROCKSDB_FIRST to HA_ERR_LAST + 1, and Oracle keeps bumping HA_ERR_LAST and we had to rebaseline it every time when we rebasing to a new release. To avoid this, I'm starting HA_ERR_ROCKSDB_FIRST to 500 and asserting it is > HA_ERR_LAST. This should probably be good enough for a few major releases without having to rebaseline, I hope. If we truly care about stability of these numbers we can bump this to a much larger number as well.

I also found out we have overlapping error numbers because one set of error uses HA_ERR_ROCKSDB_FIRST and the other set uses HA_ERR_LAST and they accidentally overlapped with each other. As a result quite a few error messages are off by 1. This fixes that. I also plans to fix this in 5.6 as well.

Reviewed By: lth

Differential Revision: D18668030

fbshipit-source-id: 014956a09fc