Fix ha_recover to support more than one transactional storage engine
Summary:
The current code assumes that one transactional storage engine (InnoDB)
and the binlog can safely rollback all prepared transactions, but with
multiple engines, it cannot. This is true, because if a transaction is
not logged to the binlog, then durable log settings are required and all
transactions should be rolled forward.

With RBR + idempotent recovery, we can allow multiple storage engines to
rollback prepared transactions.

Differential Revision: D19435700 (https://github.com/facebook/mysql-5.6/commit/4c7913bbdbad608475317ff1aeecb21567b3dc11)

fbshipit-source-id: b6dd09434d2