Make force upgrade option upgrade sysschema no matter if view count changed or not
Summary:
Current the way we decide to do fix_sys_schema is based on if there is any changes for sys schema version number, or the number of views, functions, procedures, etc.  If we only modify one of the views definition, for example, the upgrade may not really refresh the sys schema.

Normally when definitions get changed in sys schema, the sys schema version should be changed as well. But this may cause conflicts with upstream. In this case, let make force option always do the fix_sys_schema.

Reviewed By: yizhang82

Differential Revision: D20195466