Make BIT, ENUM, SET covering for MyRocks
Summary:
Today in MyRocks we have the concept of covering - meaning we can
reconstruct original value using the key (in memcmp form). For example,
integer columns in keys are covering because their memcmp representation
can be reconverted to the in-memory representation of integer fairly
easily. And there are also non-covering columns because their memcmp
representation may be "lossy" and cannot be easily converted back,
depending on the memcmp implementation.

However, there is no reason why BIT/ENUM/SET are non-covering since they
are essentially integers.

Differential Revision: D20293785 (https://github.com/facebook/mysql-5.6/commit/1b31d40a85497a706e79fe10d0533e46bad00207)

fbshipit-source-id: 9e17c111354