Acquire locks on SK unique keys during Delete
Summary:
When deleting a row in MyRocks, also acquire a lock on any unique
secondary keys. This behavior is similar to InnoDB's and reduces
conflicts resulting in failed transactions.

Some server features like dependency replication also depend on this
functionality to prevent conflicts and failing transactions. Otherwise,
it is possible for a delete, followed immediately by an insert, to cause
the insert to fail with duplicate key errors. With InnoDB, the insert
would block until the delete is committed. MyRocks should behave
similarly on delete.

Differential Revision: D20523154 (https://github.com/facebook/mysql-5.6/commit/5a888784c43d7962d50ce228d565aee1c08dea86)

fbshipit-source-id: 6471a8671bb