Call NewJemallocNodumpAllocator only if jemalloc is used
Summary:
The server fails to come up if MyRocks is built without jemalloc, but
the rocksdb_cache_dump feature is enabled. For asan/valgrind builds,
jemalloc should be disabled, but the test might still enable
rocksdb_cache_dump. Allow the asan builds to run with the feature
enabled, but generate a warning to the log instead.

Reviewed By: lloyd

Differential Revision: D20907021 (https://github.com/facebook/mysql-5.6/commit/98cdc1742b4ee516801b7835019250ae90f51d13)

fbshipit-source-id: 026c447329e