Optimize replace into to do a blind write (#1124)
Summary:
This is a diff based of https://github.com/facebook/mysql-5.6/pull/913.
The initial pull request was submitted by Zhang Yuan from Alibaba. I am
trying to continue work on this.

Approach:

During sql_insert, if the statement is a 'replace into' statement, then
we set a boolean in the lex to indicate that primary key violations can
be ignored. This is done only for tables that have an explicitly defined
primary key, does not have any triggers or secondary keys.

If the lex flag indicates that this is a blind replace into, then engine
ignores duplicate primary key errors. This in turn ensures that the sql
layer does not have to do a delete followed by a insert.

If the lex flag is set, then Write_rows_log_event sets a bit in its flag
indicating that this was an optimized 'replace into' statement
execution. This is carried over to the slave and the slave sets the lex
flag (in Write_rows_log_event::write_row()) before invoking the
handler's write_row methods.

On the slave, if blind replace bit is set for a write rows event and it
is possible to do a blind replace, we will proceed with it.

If for some reason (such as slaves having triggers or secondary keys),
we cannot do a blind replace then this will fall back to doing either an
update or a delete+insert. Note that the binlog will not have a before
image (since on master this is just an insert), but we can still do the
regular update or delete+insert.

Reference Patch: https://github.com/facebook/mysql-5.6/commit/f14c64cf950

Pull Request resolved: https://github.com/facebook/mysql-5.6/pull/1124

Reviewed By: yizhang82

Differential Revision: D21845477

fbshipit-source-id: 0617d534b47