Allow server to bootstrap with unknown variables
Summary:
When mysqld initializes for the first time, the --initialize
parameter is passed in, which puts mysqld into a bootstrapping mode.
In this mode, mysqld initializes the data directories and then exists.
For 8.0.20, plugins are no longer loaded.

However, the config file used for the normal mysqld process is also
used for the bootstrapping process. Any plugin variables specified
in the config file may result in the bootstrap process failing because
those variables now considered unknown.

To fix this, extend ignoring unknown options to include variables
during the bootstrap process.

Reviewed By: george-reynya

Differential Revision: D26284550

fbshipit-source-id: dbc6773e06d