Add coverage-src-filter option to code coverage
Summary:
This diff adds a new option **'--coverage-src-filter'** option to code coverage which limits the source files that are used to generate the code coverage metrics. The format of the option **'--coverage-src-filter=<dirname>'**.

This option can be used multiple times to specify multiple directories. All the MySQL source files (.cc and .h) from the coverage source filter directories obtained recursively (including the files in the sub-directories) will be used to generate the coverage.

This option will work in conjunction with the existing option that specifies the scope of the coverage *'--coverage-scope'*. If one or more *coverage-src-filter* options are specified then:
- If *coverage-scope* is empty or *'full'* then *coverage-src-filter* options will take the precedence.
- If *coverage-scope* is *'diff'* or *'diff:commithash'* then:
  - if source files are modified in the diff then diff will precedence and diff level coverage will be generated.
  - if no source files are found to be modified in the diff then the *coverage-src-filter* options will take precedence.

The following take summarizes the interaction between the options 'coverage-scope' and 'coverage-src-filter':
{F554900622}

The diff also creates a file 'command.log' in <build_dir>/reports/last that shows the exact mysql test command used to generate the report.

Diff D27405284 shows the changes required in ~/mysql/tools/ repos to support the new option.

Reviewed By: yizhang82

Differential Revision: D27405278

fbshipit-source-id: 231116aff42